transactionByte calculations will get its own file, first crypto moving into crypto dir
