+++ src/index.js
@@ -20,7 +20,7 @@
 import 'babel-polyfill';
 import naclFactory from 'js-nacl';
 import buffer from 'buffer';
-import crypto from './transactions/crypto';
+import crypto from './crypto';
 import dapp from './transactions/dapp';
 import delegate from './transactions/delegate';
 import multisignature from './transactions/multisignature';
