+++ src/transactions/multisignature.js
@@ -17,7 +17,7 @@
  * transactions, and signing transactions requiring multisignatures.
  * @class multisignature
  */
-import crypto from './crypto';
+import crypto from '../crypto';
 import constants from '../constants';
 import slots from '../time/slots';
 import { prepareTransaction } from './utils';
