+++ src/transactions/transfer.js
@@ -17,7 +17,7 @@
  * an individual dapp account).
  * @class transfer
  */
-import crypto from './crypto';
+import crypto from '../crypto';
 import constants from '../constants';
 import slots from '../time/slots';
 import { prepareTransaction } from './utils';
