+++ test/transactions/signature.js
@@ -13,7 +13,7 @@
  *
  */
 import signature from '../../src/transactions/signature';
-import cryptoModule from '../../src/transactions/crypto';
+import cryptoModule from '../../src/crypto';
 import slots from '../../src/time/slots';
 
 describe('signature module', () => {
