+++ test/transactions/transaction.js
@@ -14,7 +14,7 @@
  */
 import transaction from '../../src/transactions/transaction';
 import slots from '../../src/time/slots';
-import cryptoModule from '../../src/transactions/crypto';
+import cryptoModule from '../../src/crypto';
 
 describe('transaction module', () => {
 	describe('exports', () => {
