+++ test/transactions/transfer.js
@@ -13,7 +13,7 @@
  *
  */
 import transfer from '../../src/transactions/transfer';
-import cryptoModule from '../../src/transactions/crypto';
+import cryptoModule from '../../src/crypto';
 import slots from '../../src/time/slots';
 
 describe('transfer module', () => {
