+++ test/transactions/vote.js
@@ -13,7 +13,7 @@
  *
  */
 import vote from '../../src/transactions/vote';
-import cryptoModule from '../../src/transactions/crypto';
+import cryptoModule from '../../src/crypto';
 import slots from '../../src/time/slots';
 
 describe('vote module', () => {
